+++ coaster/sqlalchemy.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
 
 from __future__ import absolute_import
-import uuid
+import uuid as uuid_
 import simplejson
 from sqlalchemy import Column, Integer, DateTime, Unicode, UnicodeText, CheckConstraint, Numeric
 from sqlalchemy import event, inspect
 from sqlalchemy.sql import select, func, functions
 from sqlalchemy.types import UserDefinedType, TypeDecorator, TEXT
-from sqlalchemy.orm import composite
+from sqlalchemy.orm import composite, synonym
 from sqlalchemy.ext.declarative import declared_attr
 from sqlalchemy.ext.mutable import Mutable, MutableComposite
 from sqlalchemy.ext.hybrid import Comparator, hybrid_property
@@ -18,17 +18,17 @@ from sqlalchemy_utils.types import UUIDType
 from werkzeug.exceptions import NotFound
 from flask import Markup, url_for
 from flask_sqlalchemy import BaseQuery
-from .utils import make_name
+from .utils import make_name, uuid2buid, uuid2suuid, buid2uuid, suuid2uuid
 from .gfm import markdown
 
 
 # --- Exceptions --------------------------------------------------------------
 
-__all_exceptions = ['InvalidUuid']
+__all_exceptions = ['InvalidId']
 
 
-class InvalidUuid(NotFound):
-    """Invalid UUID. Triggers the NotFound handler when an invalid id is used in a URL."""
+class InvalidId(NotFound):
+    """Invalid id/UUID. Triggers the NotFound handler when an invalid id is used in a URL."""
     pass
 
 
@@ -60,12 +60,7 @@ def __utcnow_mssql(element, compiler, **kw):
     return 'SYSUTCDATETIME()'
 
 
-# --- Mixins ------------------------------------------------------------------
-
-__all_mixins = ['IdMixin', 'TimestampMixin', 'PermissionMixin', 'UrlForMixin',
-    'BaseMixin', 'BaseNameMixin', 'BaseScopedNameMixin', 'BaseIdNameMixin',
-    'BaseScopedIdMixin', 'BaseScopedIdNameMixin', 'CoordinatesMixin']
-
+# --- Queries and comparators -------------------------------------------------
 
 class Query(BaseQuery):
     """
@@ -78,7 +73,7 @@ class Query(BaseQuery):
         SQL EXISTS function, so the database stops counting after the first result
         is found.
         """
-        return self.session.query(self.exists()).first()[0]
+        return self.session.query(self.exists()).scalar()
 
     def isempty(self):
         """
@@ -86,20 +81,87 @@ class Query(BaseQuery):
         SQL EXISTS function, so the database stops counting after the first result
         is found.
         """
-        return not self.session.query(self.exists()).first()[0]
+        return not self.session.query(self.exists()).scalar()
 
 
-class SqlHexUuidComparator(Comparator):
+class SplitIndexComparator(Comparator):
+    """
+    Base class for comparators that support splitting a string and
+    comparing with one of the split values.
+    """
+
+    def __init__(self, expression, splitindex=None):
+        super(SplitIndexComparator, self).__init__(expression)
+        self.splitindex = splitindex
+
+
+class SqlSplitIdComparator(SplitIndexComparator):
+    """
+    Allows comparing an id value with a column, useful mostly because of
+    the splitindex feature, which splits an incoming string along the ``-``
+    character and picks one of the splits for comparison.
+    """
+    def operate(self, op, other):
+        if self.splitindex is not None and isinstance(other, basestring):
+            try:
+                other = int(other.split('-')[self.splitindex])
+            except ValueError:
+                raise InvalidId(other)
+        return op(self.__clause_element__(), other)
+
+
+class SqlHexUuidComparator(SplitIndexComparator):
     """
     Allows comparing UUID fields with hex representations of the UUID
     """
     def operate(self, op, other):
-        if not isinstance(other, uuid.UUID):
+        if not isinstance(other, uuid_.UUID):
+            if self.splitindex is not None:
+                other = other.split('-')[self.splitindex]
             try:
-                other = uuid.UUID(other)
+                other = uuid_.UUID(other)
             except ValueError:
-                raise InvalidUuid(other)
+                raise InvalidId(other)
         return op(self.__clause_element__(), other)
+
+
+class SqlBuidComparator(SplitIndexComparator):
+    """
+    Allows comparing UUID fields with URL-safe Base64 (BUID) representations
+    of the UUID
+    """
+    def operate(self, op, other):
+        if not isinstance(other, uuid_.UUID):
+            if self.splitindex is not None:
+                other = other.split('-')[self.splitindex]
+            try:
+                other = buid2uuid(other)
+            except ValueError:
+                raise InvalidId(other)
+        return op(self.__clause_element__(), other)
+
+
+class SqlSuuidComparator(SplitIndexComparator):
+    """
+    Allows comparing UUID fields with ShortUUID representations of the UUID
+    """
+    def operate(self, op, other):
+        if not isinstance(other, uuid_.UUID):
+            if self.splitindex is not None:
+                other = other.split('-')[self.splitindex]
+            try:
+                other = suuid2uuid(other)
+            except ValueError:
+                raise InvalidId(other)
+        return op(self.__clause_element__(), other)
+
+
+# --- Mixins ------------------------------------------------------------------
+
+__all_mixins = ['IdMixin', 'TimestampMixin', 'PermissionMixin', 'UrlForMixin',
+    'BaseMixin', 'BaseNameMixin', 'BaseScopedNameMixin', 'BaseIdNameMixin',
+    'BaseScopedIdMixin', 'BaseScopedIdNameMixin', 'CoordinatesMixin',
+    'UuidMixin']
 
 
 class IdMixin(object):
@@ -117,7 +179,7 @@ class IdMixin(object):
         Database identity for this model, used for foreign key references from other models
         """
         if cls.__uuid_primary_key__:
-            return Column(UUIDType(binary=False), default=uuid.uuid4, primary_key=True)
+            return Column(UUIDType(binary=False), default=uuid_.uuid4, primary_key=True)
         else:
             return Column(Integer, primary_key=True)
 
@@ -152,25 +214,106 @@ class IdMixin(object):
         return '<%s %s>' % (self.__class__.__name__, self.id)
 
 
+class UuidMixin(object):
+    """
+    Provides a ``uuid`` attribute that is either a SQL UUID column or an alias
+    to the existing ``id`` column if the class uses UUID primary keys. Also
+    provides hybrid properties ``url_id``, ``buid`` and ``suuid`` that provide
+    hex, BUID and ShortUUID representations of the ``uuid`` column.
+
+    :class:`UuidMixin` must appear before other classes in the base class order::
+
+        class MyDocument(UuidMixin, BaseMixin, db.Model):
+            pass
+
+    Compatibility table:
+
+    +-----------------------+-------------+-----------------------------------------+
+    | Base class            | Compatible? | Notes                                   |
+    +=======================+=============+=========================================+
+    | BaseMixin             | Yes         |                                         |
+    +-----------------------+-------------+-----------------------------------------+
+    | BaseIdNameMixin       | Yes         |                                         |
+    +-----------------------+-------------+-----------------------------------------+
+    | BaseNameMixin         | N/A         | ``name`` is secondary key, not ``uuid`` |
+    +-----------------------+-------------+-----------------------------------------+
+    | BaseScopedNameMixin   | N/A         | ``name`` is secondary key, not ``uuid`` |
+    +-----------------------+-------------+-----------------------------------------+
+    | BaseScopedIdMixin     | No          | Conflicting :attr:`url_id` attribute    |
+    +-----------------------+-------------+-----------------------------------------+
+    | BaseScopedIdNameMixin | No          | Conflicting :attr:`url_id` attribute    |
+    +-----------------------+-------------+-----------------------------------------+
+    """
+    @declared_attr
+    def uuid(cls):
+        """UUID column, or synonym to existing :attr:`id` column if that is a UUID"""
+        if hasattr(cls, '__uuid_primary_key__') and cls.__uuid_primary_key__:
+            return synonym('id')
+        else:
+            return Column(UUIDType(binary=False), default=uuid_.uuid4, unique=True)
+
+    @hybrid_property
+    def url_id(self):
+        """URL-friendly UUID representation as a hex string"""
+        return self.uuid.hex
+
+    @url_id.comparator
+    def url_id(cls):
+        # For some reason the test fails if we use `cls.uuid` here
+        # but works fine in the `buid` and `suuid` comparators below
+        if hasattr(cls, '__uuid_primary_key__') and cls.__uuid_primary_key__:
+            return SqlHexUuidComparator(cls.id)
+        else:
+            return SqlHexUuidComparator(cls.uuid)
+
+    @hybrid_property
+    def buid(self):
+        """URL-friendly UUID representation, using URL-safe Base64 (BUID)"""
+        return uuid2buid(self.uuid)
+
+    @buid.comparator
+    def buid(cls):
+        return SqlBuidComparator(cls.uuid)
+
+    @hybrid_property
+    def suuid(self):
+        """URL-friendly UUID representation, using ShortUUID"""
+        return uuid2suuid(self.uuid)
+
+    @suuid.comparator
+    def suuid(cls):
+        return SqlSuuidComparator(cls.uuid)
+
+
 # Supply a default value for UUID-based id columns
-def __uuid_default_listener(idcolumn):
-    @event.listens_for(idcolumn, 'init_scalar', retval=True, propagate=True)
+def __uuid_default_listener(uuidcolumn):
+    @event.listens_for(uuidcolumn, 'init_scalar', retval=True, propagate=True)
     def init_scalar(target, value, dict_):
-        value = idcolumn.columns[0].default.arg(None)
-        dict_[idcolumn.key] = value
+        value = uuidcolumn.columns[0].default.arg(None)
+        dict_[uuidcolumn.key] = value
         return value
 
 
 # Setup listeners for UUID-based subclasses
-def __configure_listener(mapper, class_):
+def __configure_id_listener(mapper, class_):
     if hasattr(class_, '__uuid_primary_key__') and class_.__uuid_primary_key__:
         __uuid_default_listener(mapper.attrs.id)
 
 
-event.listen(IdMixin, 'mapper_configured', __configure_listener, propagate=True)
+def __configure_uuid_listener(mapper, class_):
+    if hasattr(class_, '__uuid_primary_key__') and class_.__uuid_primary_key__:
+        return
+    # Only configure this listener if the class doesn't use UUID primary keys,
+    # as the `uuid` column will only be an alias for `id` in that case
+    __uuid_default_listener(mapper.attrs.uuid)
+
+
+event.listen(IdMixin, 'mapper_configured', __configure_id_listener, propagate=True)
+event.listen(UuidMixin, 'mapper_configured', __configure_uuid_listener, propagate=True)
 
 
 def make_timestamp_columns():
+    """Return two columns, created_at and updated_at, with appropriate defaults"""
     return (
         Column('created_at', DateTime, default=func.utcnow(), nullable=False),
         Column('updated_at', DateTime, default=func.utcnow(), onupdate=func.utcnow(), nullable=False),
@@ -592,9 +735,6 @@ class BaseIdNameMixin(BaseMixin):
         """The title of this object"""
         return Column(Unicode(cls.__title_length__), nullable=False)
 
-    #: The attribute containing id numbers used in the URL in id-name syntax, for external reference
-    url_id_attr = 'id'
-
     def __init__(self, *args, **kw):
         super(BaseIdNameMixin, self).__init__(*args, **kw)
         if not self.name:
@@ -608,10 +748,37 @@ class BaseIdNameMixin(BaseMixin):
         if self.title:
             self.name = unicode(make_name(self.title, maxlength=self.__name_length__))
 
-    @property
-    def url_name(self):
-        """Returns a URL name combining :attr:`url_id` and :attr:`name` in id-name syntax"""
+    @hybrid_property
+    def url_id_name(self):
+        """
+        Returns a URL name combining :attr:`url_id` and :attr:`name` in id-name
+        syntax. This property is also available as :attr:`url_name` for legacy
+        reasons.
+        """
         return '%s-%s' % (self.url_id, self.name)
+
+    @url_id_name.comparator
+    def url_id_name(cls):
+        if cls.__uuid_primary_key__:
+            return SqlHexUuidComparator(cls.id, splitindex=0)
+        elif issubclass(cls, UuidMixin):
+            return SqlHexUuidComparator(cls.uuid, splitindex=0)
+        else:
+            return SqlSplitIdComparator(cls.id, splitindex=0)
+
+    url_name = url_id_name  # Legacy name
+
+    @hybrid_property
+    def url_name_suuid(self):
+        """
+        Returns a URL name combining :attr:`name` and :attr:`suuid` in name-suuid syntax.
+        To use this, the class must derive from :class:`UuidMixin`.
+        """
+        return '%s-%s' % (self.name, self.suuid)
+
+    @url_name_suuid.comparator
+    def url_name_suuid(cls):
+        return SqlSuuidComparator(cls.uuid, splitindex=-1)
 
 
 class BaseScopedIdMixin(BaseMixin):
@@ -632,9 +799,6 @@ class BaseScopedIdMixin(BaseMixin):
     def url_id(cls):
         """Contains an id number that is unique within the parent container"""
         return Column(Integer, nullable=False)
-
-    #: The attribute containing the url id value, for external reference
-    url_id_attr = 'url_id'
 
     def __init__(self, *args, **kw):
         super(BaseScopedIdMixin, self).__init__(*args, **kw)
@@ -733,10 +897,16 @@ class BaseScopedIdNameMixin(BaseScopedIdMixin):
         if self.title:
             self.name = unicode(make_name(self.title, maxlength=self.__name_length__))
 
-    @property
-    def url_name(self):
+    @hybrid_property
+    def url_id_name(self):
         """Returns a URL name combining :attr:`url_id` and :attr:`name` in id-name syntax"""
         return '%s-%s' % (self.url_id, self.name)
+
+    @url_id_name.comparator
+    def url_id_name(cls):
+        return SqlSplitIdComparator(cls.url_id, splitindex=0)
+
+    url_name = url_id_name  # Legacy name
 
 
 class CoordinatesMixin(object):